Remove Med Dispense cdw support
